2019.5.8 - Misc, pre-release

+++ src/main/java/net/minecraft/entity/passive/EntityRabbit.java
@@ -69,7 +69,14 @@ public class EntityRabbit extends EntityAnimal
         this.jumpHelper = new EntityRabbit.RabbitJumpHelper(this);
         this.moveHelper = new EntityRabbit.RabbitMoveHelper(this);
         this.setMovementSpeed(0.0D);
+        this.initializePathFinderGoals(); // CraftBukkit - moved code
     }
+    
+    // CraftBukkit start - code from constructor
+    public void initializePathFinderGoals(){
+        this.setMovementSpeed(0.0D);
+    }
+    // CraftBukkit end
 
     protected void initEntityAI()
     {
@@ -587,11 +594,25 @@ public class EntityRabbit extends EntityAnimal
 
                         if (integer.intValue() == 0)
                         {
+                            // CraftBukkit start
+                            if (org.bukkit.craftbukkit.event.CraftEventFactory.callEntityChangeBlockEvent(this.rabbit, blockpos, Blocks.AIR, 0).isCancelled()) {
+                                return;
+                            }
+                            // CraftBukkit end
                             world.setBlockState(blockpos, Blocks.AIR.getDefaultState(), 2);
                             world.destroyBlock(blockpos, true);
                         }
                         else
                         {
+                            // CraftBukkit start
+                            if (org.bukkit.craftbukkit.event.CraftEventFactory.callEntityChangeBlockEvent(
+                                    this.rabbit,
+                                    blockpos,
+                                    block, block.getMetaFromState(iblockstate.withProperty(BlockCarrot.AGE, Integer.valueOf(integer.intValue() - 1)))
+                            ).isCancelled()) {
+                                return;
+                            }
+                            // CraftBukkit end
                             world.setBlockState(blockpos, iblockstate.withProperty(BlockCarrot.AGE, Integer.valueOf(integer.intValue() - 1)), 2);
                             world.playEvent(2001, blockpos, Block.getStateId(iblockstate));
                         }
